Extract libuv ownership primitives into acyclic uv component
Move close_ptr, proxy_ptr, and with_uv_handle from src/host/proxy.h to
a new internal header src/uv/proxy.h, preserving the asynchost
namespace and behavior. This removes the direct source dependency
http -> host caused by src/http/curl.h including src/host/proxy.h.
Update all in-tree includes to reference uv/proxy.h directly, and add
the new uv source component to scripts/source-dependencies.json with
no allowed internal CCF dependencies.
