+++ scripts/source-dependencies.json
@@ -32,6 +32,7 @@
     "tcp": ["ds"],
     "threading": [],
     "tls": ["crypto", "ds"],
-    "udp": ["ds"]
+    "udp": ["ds"],
+    "uv": []
   }
 }
