+++ src/host/tcp.h
@@ -8,8 +8,8 @@
 #include "dns.h"
 #include "ds/internal_logger.h"
 #include "ds/pending_io.h"
-#include "proxy.h"
 #include "socket.h"
+#include "uv/proxy.h"
 
 #include <netinet/in.h>
 #include <optional>
